correções de autorizão

+++ src/Controller/BarbershopsController.php
@@ -13,10 +13,6 @@ use App\Controller\AppController;
 class BarbershopsController extends AppController
 {
 
-
-
-
-
     public function isAuthorized($user){
         if ($this->request->getParam('action') === 'add') {
             return true;
@@ -56,12 +52,19 @@ class BarbershopsController extends AppController
      */
     public function view($id = null)
     {
-        $barbershop = $this->Barbershops->get($id, [
-            'contain' => ['Services']
-        ]);
+        if(is_null($id)){
+            
+            return $this->redirect(['action' => 'index']);
+        
+        }else{
 
-        $this->set('barbershop', $barbershop);
-        $this->set('_serialize', ['barbershop']);
+            $barbershop = $this->Barbershops->get($id, [
+                'contain' => ['Services']
+            ]);
+
+            $this->set('barbershop', $barbershop);
+            $this->set('_serialize', ['barbershop']);
+        }
     }
 
     /**
@@ -95,21 +98,28 @@ class BarbershopsController extends AppController
      */
     public function edit($id = null)
     {
-        $barbershop = $this->Barbershops->get($id, [
-            'contain' => []
-        ]);
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $barbershop = $this->Barbershops->patchEntity($barbershop, $this->request->getData());
-            if ($this->Barbershops->save($barbershop)) {
-                $this->Flash->success(__('The barbershop has been saved.'));
+        if(is_null($id)){
+            
+            return $this->redirect(['action' => 'index']);
+        
+        }else{
 
-                return $this->redirect(['action' => 'index']);
+            $barbershop = $this->Barbershops->get($id, [
+                'contain' => []
+            ]);
+            if ($this->request->is(['patch', 'post', 'put'])) {
+                $barbershop = $this->Barbershops->patchEntity($barbershop, $this->request->getData());
+                if ($this->Barbershops->save($barbershop)) {
+                    $this->Flash->success(__('The barbershop has been saved.'));
+
+                    return $this->redirect(['action' => 'index']);
+                }
+                $this->Flash->error(__('The barbershop could not be saved. Please, try again.'));
             }
-            $this->Flash->error(__('The barbershop could not be saved. Please, try again.'));
+            $services = $this->Barbershops->Services->find('list', ['limit' => 200]);
+            $this->set(compact('barbershop', 'services'));
+            $this->set('_serialize', ['barbershop']);
         }
-        $services = $this->Barbershops->Services->find('list', ['limit' => 200]);
-        $this->set(compact('barbershop', 'services'));
-        $this->set('_serialize', ['barbershop']);
     }
 
     /**
@@ -121,14 +131,23 @@ class BarbershopsController extends AppController
      */
     public function delete($id = null)
     {
-        $this->request->allowMethod(['post', 'delete']);
-        $barbershop = $this->Barbershops->get($id);
-        if ($this->Barbershops->delete($barbershop)) {
-            $this->Flash->success(__('The barbershop has been deleted.'));
-        } else {
-            $this->Flash->error(__('The barbershop could not be deleted. Please, try again.'));
-        }
+        if(is_null($id)){
+            
+            return $this->redirect(['action' => 'index']);
+        
+        }else{
+            if($this->request->is('get')){
+                return $this->redirect(['action' => 'index']);
+            }
+            $this->request->allowMethod(['post', 'delete']);
+            $barbershop = $this->Barbershops->get($id);
+            if ($this->Barbershops->delete($barbershop)) {
+                $this->Flash->success(__('The barbershop has been deleted.'));
+            } else {
+                $this->Flash->error(__('The barbershop could not be deleted. Please, try again.'));
+            }
 
-        return $this->redirect(['action' => 'index']);
+            return $this->redirect(['action' => 'index']);
+        }
     }
 }
